pyproject.toml
[project]
name = "viatu"
version = "0.1.0"
description = "Monitor de preços de passagens LATAM em pontos LATAM Pass"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "httpx[http2]>=0.27",
  "playwright>=1.48",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "tenacity>=9.0",
  "celery[redis]>=5.4",
  "redis>=5.2",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.2",
  "python-json-logger>=2.0",
  "tf-playwright-stealth>=1.2.0",
  "curl-cffi>=0.15.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "respx>=0.21",
  "ruff>=0.7",
  "mypy>=1.13",
  "aiosqlite>=0.20",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
